refactor error handling and nav into its own layout
