+++ client/app.js
@@ -12,7 +12,6 @@ m.route.mode = "hash"
 // m.route.mode = 'pathname'
 
 m.route(document.getElementById('app'), '/', {
-  '/'       : require('./components/Layout'),
   '/list'   : require('./components/TutorialList'),
   '/create' : require('./components/CreateTutorial'),
   '/read'   : require('./components/ReadTutorial'),
